4-aml-xgboost-deployment-samples/app-service-code/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
xgboost==1.5.0
pillow==8.3.2
numpy==1.19.5
scikit-learn==1.0.1 
azureml-core==1.35.0